Move "connection-relevant"-code from "UrlUtils" & "PageCrawler", into "ConnSupportUtils".
